separate file for catalog testing
